src/wechat_client/requirements.txt
autopy==2.1.0
baidu-aip==2.2.17.0
django-crontab==0.7.1
Pillow==6.1.0
PyQt5==5.13.0
pywin32==223
requests==2.22.0
pyperclip==1.7.0